refactor(settings_ui): lift theme-prefs handlers to _handlers_theme.py
First handler-family factory (build_theme_handlers). Validates the
factory + dict-merge assembly with stable handler-dict keys.
